IO System Rehaul
Now, there is an IO component that is stored in IOManager, ID=the
position in the list. Input is an if else block.

+++ include/GUI/OverlayManager.h
@@ -20,7 +20,7 @@ class OverlayManager : public IOBase//Overlay Manager holds overlays. Each overl
 
         void add(std::tr1::shared_ptr<leon::Panel> panel);
         leon::Panel* getTarget(const std::string& target);//returns base pointer to target
-        leon::Panel* getTarget(unsigned long long int targetID);
+        leon::Panel* getTarget(unsigned int targetID);
 
         tgui::Gui& getGui();
